单元测试bug修补

+++ lib/test/SqliteTest.php
@@ -39,7 +39,9 @@ class SqliteTest extends LibTestShell
     /* {{{ protected tearDown() */
     protected function tearDown()
     {
-        $this->dao->query('DROP TABLE test_autoincrement_and_string');
+        if (is_object($this->dao)) {
+            $this->dao->query('DROP TABLE test_autoincrement_and_string');
+        }
         if (is_file($this->file)) {
             @unlink($this->file);
             @rmdir(dirname($this->file));
@@ -151,6 +153,7 @@ class SqliteTest extends LibTestShell
             $this->dao->clear()->insert(array('sign' => 'begin'))->affectedRows(),
             1, 'Affected Rows MUST BE 1 before commit.'
         );
+        //TODO: 这里应该用另外一个连接来检查
         $this->dao->commit();
         $this->assertEquals(
             $this->dao->clear()->select('COUNT(*)')->getOne(),
